removed unncessary extensions renamed the extension file to daggerExtension for better differentiation
